Working progress

+++ backend/src/main/java/com/danielfreitassc/backend/services/ReservationService.java
@@ -17,8 +17,11 @@ import com.danielfreitassc.backend.dtos.ReservationResponseDto;
 import com.danielfreitassc.backend.mappers.ReservationMapper;
 import com.danielfreitassc.backend.models.ReservationEntity;
 import com.danielfreitassc.backend.models.RoomEntity;
+import com.danielfreitassc.backend.models.RoomStatus;
+import com.danielfreitassc.backend.models.UserEntity;
 import com.danielfreitassc.backend.repositories.ReservationRepository;
 import com.danielfreitassc.backend.repositories.RoomRepository;
+import com.danielfreitassc.backend.repositories.UserRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -28,10 +31,16 @@ public class ReservationService {
     private final ReservationRepository reservationRepository;
     private final ReservationMapper reservationMapper;
     private final RoomRepository roomRepository;
+    private final UserRepository userRepository;
 
     public ReservationResponseDto create(ReservationRequestDto reservationRequestDto) {
         Optional<RoomEntity> room = roomRepository.findById(reservationRequestDto.roomId());
         if(room.isEmpty()) throw  new ResponseStatusException(HttpStatus.NOT_FOUND,"Quarto não encontrado");
+
+        Optional<UserEntity> user = userRepository.findById(reservationRequestDto.userId());
+        if(user.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND,"Usuário não encontrado");
+
+        if(room.get().getStatus() == RoomStatus.RESERVADO) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Quarto já reservado");
         if (reservationRequestDto.endTime().isBefore(reservationRequestDto.startTime())) throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "End time deve ser posterior ao start time");
 
         long durationInMinutes = Duration.between(reservationRequestDto.startTime(), reservationRequestDto.endTime()).toMinutes();
@@ -49,6 +58,10 @@ public class ReservationService {
 
         ReservationEntity reservation = reservationMapper.toEntity(reservationRequestDto);
         reservation.setTotalCost(totalCost);
+        RoomEntity roomEntity = room.get();
+        roomEntity.setStatus(RoomStatus.RESERVADO);
+        roomRepository.save(roomEntity);
+        reservation.setStatus(RoomStatus.RESERVADO);
         reservationRepository.save(reservation);
         return  reservationMapper.toDto(reservation);
     }
@@ -64,17 +77,71 @@ public class ReservationService {
         return  reservationMapper.toDto(reservation.get());
     }
 
-    public ReservationResponseDto update(UUID id,ReservationRequestDto reservationRequestDto) {
+    public ReservationResponseDto update(UUID id, ReservationRequestDto reservationRequestDto) {
         Optional<ReservationEntity> reservation = reservationRepository.findById(id);
-        if(reservation.isEmpty()) throw  new ResponseStatusException(HttpStatus.NOT_FOUND,"Nenhuma reserva encontrada");
+        if (reservation.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Nenhuma reserva encontrada");
+    
+        if (reservationRequestDto.endTime().isBefore(reservationRequestDto.startTime())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "End time deve ser posterior ao start time");
+        }
+    
+        Optional<RoomEntity> newRoom = roomRepository.findById(reservationRequestDto.roomId());
+        if (newRoom.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Quarto não encontrado");
+        
+        Optional<UserEntity> user = userRepository.findById(reservation.get().getUserEntity().getId());
+        if(user.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND,"Usuário não encontrado");
+
+        RoomEntity currentRoom = reservation.get().getRoomEntity();
+    
+        if (!currentRoom.getId().equals(reservationRequestDto.roomId())) {
+            currentRoom.setStatus(RoomStatus.DISPONIVEL);
+            roomRepository.save(currentRoom);
+            RoomEntity newRoomEntity = newRoom.get();
+    
+            if (newRoomEntity.getStatus() == RoomStatus.RESERVADO) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "O quarto já está reservado");
+            }
+    
+            newRoomEntity.setStatus(RoomStatus.RESERVADO);
+            roomRepository.save(newRoomEntity);
+        }
+    
+        long durationInMinutes = Duration.between(reservationRequestDto.startTime(), reservationRequestDto.endTime()).toMinutes();
+        if (durationInMinutes <= 0) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Duração inválida da reserva");
+        }
+    
+        BigDecimal durationInHours = BigDecimal.valueOf(durationInMinutes).divide(BigDecimal.valueOf(60), 2, RoundingMode.HALF_UP);
+    
+        BigDecimal roomPrice = newRoom.get().getPrice();
+        if (roomPrice == null || roomPrice.compareTo(BigDecimal.ZERO) == 0) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Preço do quarto inválido");
+        }
+    
+        BigDecimal totalCost = roomPrice.multiply(durationInHours);
+    
         ReservationEntity reservationEntity = reservationMapper.toEntity(reservationRequestDto);
         reservationEntity.setId(id);
-        return  reservationMapper.toDto(reservationRepository.save(reservationEntity));
+        reservationEntity.setStatus(RoomStatus.RESERVADO);
+        reservationEntity.setTotalCost(totalCost);
+    
+        return reservationMapper.toDto(reservationRepository.save(reservationEntity));
     }
+    
 
     public ReservationResponseDto delete(UUID id) {
         Optional<ReservationEntity> reservation = reservationRepository.findById(id);
         if(reservation.isEmpty()) throw  new ResponseStatusException(HttpStatus.NOT_FOUND,"Nenhuma reserva encontrada");
+
+        Optional<UserEntity> user = userRepository.findById(reservation.get().getUserEntity().getId());
+        if(user.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND,"Usuário não encontrado");
+
+        Optional<RoomEntity> room = roomRepository.findById(reservation.get().getRoomEntity().getId());
+        if(room.isEmpty()) throw  new ResponseStatusException(HttpStatus.NOT_FOUND,"Quarto não encontrado");
+        
+        RoomEntity roomEntity = room.get();
+        roomEntity.setStatus(RoomStatus.DISPONIVEL);
+        roomRepository.save(roomEntity);
         reservationRepository.delete(reservation.get());
         return  reservationMapper.toDto(reservation.get());
     }
